+++ Task_1_1_3/src/main/java/ru/nsu/laptev/Sub.java
@@ -48,7 +48,7 @@ public class Sub extends Expression {
      * @return the value of subtraction.
      */
     @Override
-    public int eval(String exp) {
+    public double eval(String exp) {
         return (demunitive.eval(exp) / deductible.eval(exp));
     }
 }
